Extract CafeModel into individual file
