+++ src/ViewModels/CafeModel.ts
@@ -1,0 +1,14 @@
+export type CafeModel = {
+  name: string,
+  email: string,
+  imageLink: string,
+  street: string,
+  city: string,
+  zipCode: string,
+  country: string,
+  state: string,
+  phone: string,
+  dataAddress: string,
+  countryWithState: string,
+  location: string,
+}
